requirements.txt
requests
matplotlib
numpy
python-dotenv
midea-inventor-lib
hashlib
flask
sqlalchemy
flask-cors
pymysql